+++ pkg/cloud/services/route53/errors.go
@@ -37,24 +37,6 @@ var rateLimitHitError = &microerror.Error{
 	Kind: "throttlingRateExceededError",
 }
 
-// IsIngressNotRead asserts ingressNotReadyError.
-func IsIngressNotReady(err error) bool {
-	return microerror.Cause(err) == ingressNotReadyError
-}
-
-var ingressNotReadyError = &microerror.Error{
-	Kind: "ingressNotReadyError",
-}
-
-// IsTooManyICServices asserts tooManyICServicesError.
-func IsTooManyICServices(err error) bool {
-	return microerror.Cause(err) == tooManyICServicesError
-}
-
-var tooManyICServicesError = &microerror.Error{
-	Kind: "tooManyICServicesError",
-}
-
 func wrapRoute53Error(err error) error {
 	if code, ok := awserrors.Code(errors.Cause(err)); ok {
 		switch code {